+++ Runtime/MultiSceneLoader.cs
@@ -15,7 +15,6 @@
 // * You should have received a copy of the GNU General Public License
 // * along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.Events;
@@ -35,7 +34,8 @@ namespace HH.MultiSceneTools
         public static UnityEvent<SceneCollection, collectionLoadMode> OnSceneCollectionLoadDebug = new UnityEvent<SceneCollection, collectionLoadMode>();
         public static int getDebugEventCount {get; private set;}
         private static bool IsLoggingOnSceneLoad;
-        static SceneCollection currentlyLoaded;
+        private static Scene loadedBootScene;
+        public static SceneCollection currentlyLoaded {private set; get;}
         public static string getLoadedCollectionTitle => currentlyLoaded.Title;
 
         #if UNITY_EDITOR
@@ -86,7 +86,48 @@ namespace HH.MultiSceneTools
             }
             OnSceneCollectionLoadDebug?.Invoke(TargetCollection, mode);
             OnSceneCollectionLoaded?.Invoke(TargetCollection, mode);
+
+            #if UNITY_EDITOR
             MultiSceneToolsConfig.instance.setCurrCollection(currentlyLoaded);
+            #endif
+        }
+
+        public static void loadCollection(SceneCollection Collection, collectionLoadMode mode)
+        {
+            if(MultiSceneToolsConfig.instance.LogOnSceneChange)
+                AddLogOnLoad();
+
+            if(Collection == null)
+            {
+                throw new System.NullReferenceException();
+            }
+
+            if(Collection.SceneNames.Count == 0)
+            {
+                Debug.LogWarning("Attempted to load a scene collection that contains no scenes", Collection);
+                return;
+            }
+
+            switch(mode)
+            {
+                case collectionLoadMode.Difference:
+                    loadDifference(Collection);
+                    break;
+
+                case collectionLoadMode.Replace:
+                    loadReplace(Collection);
+                    break;
+
+                case collectionLoadMode.Additive:
+                    loadAdditive(Collection);
+                    break;
+            }
+            OnSceneCollectionLoadDebug?.Invoke(Collection, mode);
+            OnSceneCollectionLoaded?.Invoke(Collection, mode);
+            
+            #if UNITY_EDITOR
+            MultiSceneToolsConfig.instance.setCurrCollection(currentlyLoaded);
+            #endif
         }
 
         static void loadDifference(SceneCollection Collection)
@@ -96,20 +137,24 @@ namespace HH.MultiSceneTools
                 throw new UnityException("No currently loaded scene collection.");
             }
 
-            bool loadBoot = MultiSceneToolsConfig.instance.UseBootScene;
             string bootScene = getBootSceneName();
-            bool hasBootScene = false;
+            bool shouldKeepBoot = false;
             bool shouldReplaceScene = false;
             
+            if(loadedBootScene.name != null)
+                shouldKeepBoot = true;
+
+            if(currentlyLoaded.SceneNames.Contains(bootScene) && MultiSceneToolsConfig.instance.UseBootScene)
+            {
+                shouldKeepBoot = true;
+                loadedBootScene = MultiSceneToolsConfig.instance.BootScene;
+            }
+
             // Unload Differences
+            int unloadedScenes = 0;
             for (int i = 0; i < currentlyLoaded.SceneNames.Count; i++)
             {
                 bool difference = true;
-                if(currentlyLoaded.SceneNames[i] == bootScene && loadBoot)
-                {
-                    hasBootScene = true;
-                    continue;
-                }
                 foreach (string targetScene in Collection.SceneNames)
                 {
                     if(currentlyLoaded.SceneNames[i].Equals(targetScene))
@@ -120,11 +165,17 @@ namespace HH.MultiSceneTools
                 if(!difference)
                     continue;
                 
-                if(i != currentlyLoaded.SceneNames.Count-1)
+                if(currentlyLoaded.SceneNames[i] == bootScene && shouldKeepBoot)
+                    continue;
+
+                if(unloadedScenes != currentlyLoaded.SceneNames.Count-1 || loadedBootScene.name != null)
+                {
+                    unloadedScenes++;
                     unload(currentlyLoaded.SceneNames[i]);
+                }
                 else
                 {
-                    if(!hasBootScene)
+                    if(!shouldKeepBoot)
                         shouldReplaceScene = true;
                     break;
                 }
@@ -135,6 +186,9 @@ namespace HH.MultiSceneTools
                 bool difference = true;
                 foreach (string LoadedScene in currentlyLoaded.SceneNames)
                 {
+                    if(targetScene.Equals(bootScene) && loadedBootScene.name != null)
+                        difference = false;
+                    
                     if(targetScene.Equals(LoadedScene))
                     {
                         difference = false;
@@ -149,47 +203,68 @@ namespace HH.MultiSceneTools
                 }
             }
             currentlyLoaded = Collection;
-            MultiSceneToolsConfig.instance.setCurrCollection(currentlyLoaded);
         }
 
         static void loadReplace(SceneCollection Collection)
         {
             bool loadBoot = MultiSceneToolsConfig.instance.UseBootScene;
             string bootScene = getBootSceneName();
-            bool hasBootScene = false;
-
+            bool shouldKeepBoot = false;
             bool shouldReplaceScene = false;
+
+            if(loadedBootScene.name != null)
+                shouldKeepBoot = true;
+
+            if(currentlyLoaded.SceneNames.Contains(bootScene) && loadBoot)
+            {
+                shouldKeepBoot = true;
+                loadedBootScene = MultiSceneToolsConfig.instance.BootScene;
+            }
+
+            if(loadBoot && loadedBootScene.name == null)
+                shouldReplaceScene = true;
+
+            // Unload Scenes
+            int unloadedScenes = 0;
             for (int i = 0; i < currentlyLoaded.SceneNames.Count; i++)
             {
-                if(currentlyLoaded.SceneNames[i] == bootScene && loadBoot)
+                if(shouldReplaceScene)
+                    break;
+
+                if(currentlyLoaded.SceneNames.Count < 2 && !loadBoot)
                 {
-                    hasBootScene = true;
+                    shouldReplaceScene = true;
                     continue;
                 }
-                if(i != currentlyLoaded.SceneNames.Count-1)
+
+                if(currentlyLoaded.SceneNames[i].Equals(bootScene) && loadedBootScene.name != null)
+                    continue;
+
+                if(unloadedScenes != currentlyLoaded.SceneNames.Count-1 || loadedBootScene.name != null)
+                {
+                    unloadedScenes++;
                     unload(currentlyLoaded.SceneNames[i]);
+                }
                 else
                 {
-                    if(!hasBootScene)
+                    if(!shouldKeepBoot)
                         shouldReplaceScene = true;
                     break;
                 }
             }
 
-            if(!hasBootScene && loadBoot)
-            {
-                load(bootScene, LoadSceneMode.Additive);
-            }
-
             for (int i = 0; i < Collection.SceneNames.Count; i++)
             {
-                if(loadBoot) // ! problem code, consider reworking the boot scene functions
+                if(loadBoot)
                 {
                     if(Collection.SceneNames[i] == bootScene)
                         continue;
 
                     if(shouldReplaceScene)
+                    {
                         load(Collection.SceneNames[i], LoadSceneMode.Single);
+                        shouldReplaceScene = false;
+                    }
                     else
                         load(Collection.SceneNames[i], LoadSceneMode.Additive);
                 }
@@ -199,16 +274,18 @@ namespace HH.MultiSceneTools
                     load(Collection.SceneNames[i], LoadSceneMode.Additive);
             }
             currentlyLoaded = Collection;
-            MultiSceneToolsConfig.instance.setCurrCollection(currentlyLoaded);
         }
+
 
         static void loadAdditive(SceneCollection Collection)
         {
-            for (int i = 0; i < Collection.SceneNames.Count; i++)
-            {
-                load(Collection.SceneNames[i], LoadSceneMode.Additive);
-            }
-            MultiSceneToolsConfig.instance.setCurrCollection(currentlyLoaded);
+            throw new System.NotImplementedException();
+            
+            // for (int i = 0; i < Collection.SceneNames.Count; i++)
+            // {
+            //     load(Collection.SceneNames[i], LoadSceneMode.Additive);
+            // }
+            // MultiSceneToolsConfig.instance.setCurrCollection(currentlyLoaded);
         }
 
         static SceneCollection FindCollection(string CollectionTitle)
@@ -224,10 +301,7 @@ namespace HH.MultiSceneTools
 
         static string getBootSceneName()
         {
-            string bootPath = MultiSceneToolsConfig.instance._BootScenePath;
-            string[] split = bootPath.Split('/');
-            string bootName = split[split.Length-1];
-            return bootName.Split('.')[0];
+            return MultiSceneToolsConfig.instance.BootScene.name;
         }
 
         static void unload(string SceneName)